Share "find_postgres_bin" logic
