+++ crates/bestool/src/actions/tamanu.rs
@@ -125,44 +125,84 @@ pub fn find_existing_version() -> Result<Version> {
 		.version)
 }
 
-#[cfg(all(feature = "tamanu-pg-common", not(windows)))]
-fn find_postgres_bin(name: &str) -> Result<OsString> {
-	Ok(name.into())
-}
+#[cfg(feature = "tamanu-pg-common")]
+pub fn find_postgres_bin(name: &str) -> Result<OsString> {
+	use std::env;
 
-#[cfg(all(feature = "tamanu-pg-common", windows))]
-#[tracing::instrument(level = "debug")]
-fn find_postgres_bin(name: &str) -> Result<OsString> {
+	#[allow(dead_code)]
+	#[tracing::instrument(level = "debug")]
+	fn find_from_installation(root: &str, name: &str) -> Result<OsString> {
+		let version = fs::read_dir(root)
+			.into_diagnostic()?
+			.filter_map(|res| {
+				res.map(|dir| {
+					dir.file_name()
+						.into_string()
+						.ok()
+						.filter(|name| name.parse::<u32>().is_ok())
+				})
+				.transpose()
+			})
+			// Use `u32::MAX` in case of `Err` so that we always catch IO errors.
+			.max_by_key(|res| {
+				res.as_ref()
+					.cloned()
+					.map(|n| n.parse::<u32>().unwrap())
+					.unwrap_or(u32::MAX)
+			})
+			.ok_or_else(|| miette!("the Postgres root {root} is empty"))?
+			.into_diagnostic()?;
+
+		let exec_file_name = if cfg!(windows) {
+			format!("{name}.exe")
+		} else {
+			format!("{name}")
+		};
+		Ok([root, version.as_str(), "bin", &exec_file_name]
+			.iter()
+			.collect::<PathBuf>()
+			.into())
+	}
+
+	#[allow(dead_code)]
+	fn is_in_path(name: &str) -> Option<PathBuf> {
+		let var = env::var_os("PATH")?;
+
+		// Separate PATH value into paths
+		let paths_iter = env::split_paths(&var);
+
+		// Attempt to read each path as a directory
+		let dirs_iter = paths_iter.filter_map(|path| fs::read_dir(path).ok());
+
+		for dir in dirs_iter {
+			let mut matches_iter = dir
+				.filter_map(|file| file.ok())
+				.filter(|file| file.file_name() == name);
+			if let Some(file) = matches_iter.next() {
+				return Some(file.path());
+			}
+		}
+
+		None
+	}
+
 	// On Windows, find `psql` assuming the standard installation using the installer
 	// because PATH on Windows is not reliable.
 	// See https://github.com/rust-lang/rust/issues/37519
-	let root = r"C:\Program Files\PostgreSQL";
-	let version = fs::read_dir(root)
-		.into_diagnostic()?
-		.inspect(|res| tracing::debug!(?res, "reading PostgreSQL installation"))
-		.filter_map(|res| {
-			res.map(|dir| {
-				dir.file_name()
-					.into_string()
-					.ok()
-					.filter(|name| name.parse::<u32>().is_ok())
-			})
-			.transpose()
-		})
-		// Use `u32::MAX` in case of `Err` so that we always catch IO errors.
-		.max_by_key(|res| {
-			res.as_ref()
-				.cloned()
-				.map(|n| n.parse::<u32>().unwrap())
-				.unwrap_or(u32::MAX)
-		})
-		.ok_or_else(|| miette!("the Postgres root {root} is empty"))?
-		.into_diagnostic()?;
+	#[cfg(windows)]
+	return find_from_installation(r"C:\Program Files\PostgreSQL", name);
 
-	Ok([root, version.as_str(), "bin", &format!("{name}.exe")]
-		.iter()
-		.collect::<PathBuf>()
-		.into())
+	#[cfg(target_os = "linux")]
+	if is_in_path(name).is_some() {
+		return Ok(name.into());
+	} else {
+		// Ubuntu reccomends to use pg_ctlcluster over pg_ctl and doesn't put pg_ctl in PATH.
+		// Still, it should be fine for temporary database.
+		return find_from_installation(r"/usr/lib/postgresql", name);
+	}
+
+	#[cfg(not(any(windows, target_os = "linux")))]
+	return Ok(name.into());
 }
 
 #[cfg(feature = "tamanu-pg-common")]
