+++ crates/bestool/src/actions/tamanu/roots.rs
@@ -1,4 +1,7 @@
-use std::{path::{Path, PathBuf}, sync::LazyLock};
+use std::{
+	path::{Path, PathBuf},
+	sync::LazyLock,
+};
 
 use leon::Template;
 use leon_macros::template;
@@ -50,7 +53,8 @@ pub fn version_of_root(root: impl AsRef<Path> + std::fmt::Debug) -> Result<Optio
 	let root = root.as_ref();
 
 	if let Some(name) = root.file_name().and_then(|name| name.to_str()) {
-		static RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(release-)?v?(\d+\.\d+\.\d+)($|/)").unwrap());
+		static RE: LazyLock<Regex> =
+			LazyLock::new(|| Regex::new(r"(release-)?v?(\d+\.\d+\.\d+)($|/)").unwrap());
 		if let Some(ver) = RE.find(name) {
 			if let Ok(semver) = Version::parse(ver.as_str()) {
 				trace!(?semver, "parsed version from path");
